% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deltaMeanLogistic}
\alias{deltaMeanLogistic}
\title{Standard Error: Mean of Logistic Regression}
\usage{
deltaMeanLogistic(x, y, cov)
}
\arguments{
\item{x}{covariates}

\item{y}{outcomes}

\item{cov}{covariance matrix of coefficients}
}
\description{
Standard Error: Mean of Logistic Regression
}
