% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deltaMeanOLS}
\alias{deltaMeanOLS}
\title{Standard Error: Mean of OLS Regression}
\usage{
deltaMeanOLS(x, y, cov)
}
\arguments{
\item{x}{covariates}

\item{y}{outcomes}

\item{cov}{covariance matrix of coefficients}
}
\description{
Standard Error: Mean of OLS Regression
}
