% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_helpers.R
\name{df_time}
\alias{df_time}
\title{Survival Data Frame Time Variable Expansion}
\usage{
df_time(df, t_current)
}
\arguments{
\item{df}{data frame}

\item{t_current}{current time}
}
\description{
Survival Data Frame Time Variable Expansion
}
