% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_fit_site.R
\name{LF_fit_site}
\alias{LF_fit_site}
\title{Likelihood Factor Estimated from Data using sl3.}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LF_base} object
}
\description{
Uses an \code{sl3} learner to estimate a likelihood factor from data.
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
}
\section{Constructor}{

  \code{define_lf(LF_fit, name, learner, ..., type = "density")}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{learner}}{An sl3 learner to be used to estimate the factor
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{site}}{Value that indicates an observation being a trial data
    }
    \item{\code{site_node}}{Node of the trial indicator
    }
    \item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{learner}}{The learner or learner fit object}
    }
}

\concept{Likelihood objects}
\keyword{data}
\section{Super classes}{
\code{\link[tmle3:LF_base]{tmle3::LF_base}} -> \code{\link[tmle3:LF_fit]{tmle3::LF_fit}} -> \code{LF_fit_site}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LF_fit_site$new()}}
\item \href{#method-delayed_train}{\code{LF_fit_site$delayed_train()}}
\item \href{#method-clone}{\code{LF_fit_site$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="cf_values">}\href{../../tmle3/html/LF_base.html#method-cf_values}{\code{tmle3::LF_base$cf_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="get_likelihood">}\href{../../tmle3/html/LF_base.html#method-get_likelihood}{\code{tmle3::LF_base$get_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="print">}\href{../../tmle3/html/LF_base.html#method-print}{\code{tmle3::LF_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_fit" data-id="get_density">}\href{../../tmle3/html/LF_fit.html#method-get_density}{\code{tmle3::LF_fit$get_density()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_fit" data-id="get_mean">}\href{../../tmle3/html/LF_fit.html#method-get_mean}{\code{tmle3::LF_fit$get_mean()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_fit" data-id="train">}\href{../../tmle3/html/LF_fit.html#method-train}{\code{tmle3::LF_fit$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_site$new(
  name,
  learner,
  ...,
  site = 1,
  site_node = "A",
  type = "density"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delayed_train"></a>}}
\if{latex}{\out{\hypertarget{method-delayed_train}{}}}
\subsection{Method \code{delayed_train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_site$delayed_train(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_site$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
