% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_helpers.R
\name{long_to_mat}
\alias{long_to_mat}
\title{Survival Data Vector to Matrix}
\usage{
long_to_mat(x, id, time)
}
\arguments{
\item{x}{data vector}

\item{id}{patient id}

\item{time}{observation time}
}
\description{
Survival Data Vector to Matrix
}
