% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_SCT.R
\name{Param_SCT}
\alias{Param_SCT}
\title{Survival Curve}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
Survival Curve

Survival Curve
}
\section{Constructor}{

  \code{define_param(Param_survival, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this treatment
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention
    }
}
}

\seealso{
Other Parameters: 
\code{\link{Param_AET}}
}
\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_SCT}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_SCT$new()}}
\item \href{#method-long_to_mat}{\code{Param_SCT$long_to_mat()}}
\item \href{#method-hm_to_sm}{\code{Param_SCT$hm_to_sm()}}
\item \href{#method-clever_covariates_internal}{\code{Param_SCT$clever_covariates_internal()}}
\item \href{#method-clever_covariates}{\code{Param_SCT$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_SCT$estimates()}}
\item \href{#method-clone}{\code{Param_SCT$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$new(
  observed_likelihood,
  target_times = NULL,
  onsite = 1,
  offsite = 0,
  ...,
  covariate_node = "W",
  site_node = "S",
  outcome_node = "N"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-long_to_mat"></a>}}
\subsection{Method \code{long_to_mat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$long_to_mat(x, id, time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hm_to_sm"></a>}}
\subsection{Method \code{hm_to_sm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$hm_to_sm(hm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates_internal"></a>}}
\subsection{Method \code{clever_covariates_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$clever_covariates_internal(
  tmle_task = NULL,
  fold_number = "full",
  subset_times = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$clever_covariates(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SCT$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
