% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_SET.R
\name{Param_SET}
\alias{Param_SET}
\title{Survival Outcome Transportation}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Survival Outcome Transportation

Survival Outcome Transportation
}
\section{Constructor}{

  \code{define_param(Param_survival, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this treatment
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention
    }
}
}

\seealso{
Other Parameters: 
\code{\link{Param_AET}},
\code{\link{Param_AOT}},
\code{\link{Param_SOT}}
}
\concept{Parameters}
\keyword{data}
\section{Super classes}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{\link[tmle3trans:Param_AOT]{tmle3trans::Param_AOT}} -> \code{\link[tmle3trans:Param_SOT]{tmle3trans::Param_SOT}} -> \code{Param_SET}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_SET$new()}}
\item \href{#method-clever_covariates_internal}{\code{Param_SET$clever_covariates_internal()}}
\item \href{#method-clone}{\code{Param_SET$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3trans" data-topic="Param_SOT" data-id="clever_covariates">}\href{../../tmle3trans/html/Param_SOT.html#method-clever_covariates}{\code{tmle3trans::Param_SOT$clever_covariates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3trans" data-topic="Param_SOT" data-id="estimates">}\href{../../tmle3trans/html/Param_SOT.html#method-estimates}{\code{tmle3trans::Param_SOT$estimates()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SET$new(
  observed_likelihood,
  intervention,
  onsite = 1,
  offsite = 0,
  target_times = NULL,
  fit_s_marginal = "empirical",
  ...,
  outcome_node = "failed"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates_internal"></a>}}
\if{latex}{\out{\hypertarget{method-clever_covariates_internal}{}}}
\subsection{Method \code{clever_covariates_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SET$clever_covariates_internal(
  tmle_task = NULL,
  fold_number = "full",
  subset_times = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_SET$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
