% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_TR.R
\name{Param_TR}
\alias{Param_TR}
\title{Average Treatment Effect}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
Parameter definition for the Average Treatment Effect (ATE).
}
\section{Constructor}{

  \code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
    }
    \item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
    }
    \item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
    }
    \item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
    }
    \item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
    }
}
}

\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_TR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_TR$new()}}
\item \href{#method-clever_covariates}{\code{Param_TR$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_TR$estimates()}}
\item \href{#method-clone}{\code{Param_TR$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TR$new(
  observed_likelihood,
  onsite = 1,
  offsite = 0,
  ...,
  covariate_node = "W",
  site_node = "A",
  outcome_node = "Y"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TR$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TR$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_TR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
