% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_AET.R
\name{tmle_AET}
\alias{tmle_AET}
\title{All Treatment Specific Means}
\usage{
tmle_AET(onsite = 1, offsite = 0, fit_s_marginal = "empirical", ...)
}
\arguments{
\item{onsite}{value for onsite}

\item{offsite}{value for offsite}
}
\description{
O=(W,S,A,Y)
W=Covariates
S=Location
A=Intervention
Y=Outcome (binary or bounded continuous)
}
