% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_AOT.R
\name{tmle_AOT}
\alias{tmle_AOT}
\title{All Treatment Specific Means}
\usage{
tmle_AOT(onsite = 1, offsite = 0, fit_s_marginal = "empirical", ...)
}
\arguments{
\item{onsite}{value for onsite}

\item{offsite}{value for offsite}
}
\description{
O=(W,S,Y)
W=Covariates
S=Location
Y=Outcome (binary or bounded continuous)
}
