% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_SCT.R
\name{tmle_SCT}
\alias{tmle_SCT}
\title{All Treatment Specific Means}
\usage{
tmle_SCT(target_times = NULL, onsite = 1, offsite = 0, ...)
}
\arguments{
\item{onsite}{value for onsite}

\item{offsite}{value for offsite}
}
\description{
All Treatment Specific Means
}
