% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_TR.R
\name{tmle_TR}
\alias{tmle_TR}
\title{All Treatment Specific Means}
\usage{
tmle_TR(onsite = 1, offsite = 0, ...)
}
\arguments{
\item{modification}{s(W)}

\item{mapping}{f(A, modification)}

\item{n_samples}{number of samples to draw for each observation}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (continuous)
Y=Outcome (binary or bounded continuous)
}
