% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_helpers.R
\name{transform_data}
\alias{transform_data}
\title{Survival Data All Time Expansion}
\usage{
transform_data(data, node_list)
}
\arguments{
\item{data}{data}

\item{node_list}{node_list}
}
\description{
Survival Data All Time Expansion
}
