% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_SET.R
\name{tmle_SET}
\alias{tmle_SET}
\title{Wrapper}
\usage{
tmle_SET(
  intervention,
  onsite = 1,
  offsite = 0,
  target_times = NULL,
  fit_s_marginal = "empirical",
  ...
)
}
\arguments{
\item{onsite}{value for onsite}

\item{offsite}{value for offsite}
}
\description{
Wrapper
}
